requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.txt
#
appnope==0.1.2
    # via
    #   -r requirements.txt
    #   ipython
asgiref==3.3.4
    # via
    #   -r requirements.txt
    #   django
backcall==0.2.0
    # via
    #   -r requirements.txt
    #   ipython
certifi==2021.5.30
    # via
    #   -r requirements.txt
    #   requests
cffi==1.14.5
    # via
    #   -r requirements.txt
    #   cryptography
chardet==4.0.0
    # via
    #   -r requirements.txt
    #   requests
click==8.0.1
    # via
    #   -r requirements.txt
    #   pip-tools
cryptography==3.4.7
    # via
    #   -r requirements.txt
    #   jwcrypto
decorator==5.0.9
    # via
    #   -r requirements.txt
    #   ipython
deprecated==1.2.12
    # via
    #   -r requirements.txt
    #   jwcrypto
django==3.2.4
    # via
    #   -r requirements.txt
    #   django-filter
    #   django-oauth-toolkit
    #   djangorestframework
django-filter==2.4.0
    # via -r requirements.txt
django-oauth-toolkit==1.5.0
    # via -r requirements.txt
djangorestframework==3.12.4
    # via -r requirements.txt
idna==2.10
    # via
    #   -r requirements.txt
    #   requests
ipython==7.24.1
    # via -r requirements.txt
ipython-genutils==0.2.0
    # via
    #   -r requirements.txt
    #   traitlets
jedi==0.18.0
    # via
    #   -r requirements.txt
    #   ipython
jwcrypto==0.9.1
    # via
    #   -r requirements.txt
    #   django-oauth-toolkit
markdown==3.3.4
    # via -r requirements.txt
matplotlib-inline==0.1.2
    # via
    #   -r requirements.txt
    #   ipython
oauthlib==3.1.1
    # via
    #   -r requirements.txt
    #   django-oauth-toolkit
parso==0.8.2
    # via
    #   -r requirements.txt
    #   jedi
pep517==0.10.0
    # via
    #   -r requirements.txt
    #   pip-tools
pexpect==4.8.0
    # via
    #   -r requirements.txt
    #   ipython
pickleshare==0.7.5
    # via
    #   -r requirements.txt
    #   ipython
pip-tools==6.2.0
    # via -r requirements.txt
prompt-toolkit==3.0.19
    # via
    #   -r requirements.txt
    #   ipython
ptyprocess==0.7.0
    # via
    #   -r requirements.txt
    #   pexpect
pycparser==2.20
    # via
    #   -r requirements.txt
    #   cffi
pygments==2.9.0
    # via
    #   -r requirements.txt
    #   ipython
pytz==2021.1
    # via
    #   -r requirements.txt
    #   django
requests==2.25.1
    # via
    #   -r requirements.txt
    #   django-oauth-toolkit
six==1.16.0
    # via
    #   -r requirements.txt
    #   django-oauth-toolkit
    #   jwcrypto
sqlparse==0.4.1
    # via
    #   -r requirements.txt
    #   django
toml==0.10.2
    # via
    #   -r requirements.txt
    #   pep517
traitlets==5.0.5
    # via
    #   -r requirements.txt
    #   ipython
    #   matplotlib-inline
urllib3==1.26.5
    # via
    #   -r requirements.txt
    #   requests
wcwidth==0.2.5
    # via
    #   -r requirements.txt
    #   prompt-toolkit
wheel==0.36.2
    # via pip-tools
wrapt==1.12.1
    # via
    #   -r requirements.txt
    #   deprecated

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
